ci(mypy): expand blocking allowlist 10 -> 29 modules
Ratchet the blocking mypy gate from the 10 security/data-governance core
modules to 29 by making the memory-management surface type-clean.

Added (all light annotation only -- no logic changes):
- no-any-return wraps: memory_importance_scorer, memory_reranker,
  memory_provenance, memory_tier_manager, memory_versioner, plugin_loader,
  undo_manager, notification_manager, webhook_manager
- generic type params (Dict/List/Set/Tuple): graph_compactor,
  memory_consolidator, memory_deduplication, cross_agent_sharing,
  memory_confidence
- missing arg/return annotations: memory_management, structured_memory,
  security_mcp, cross_agent_sharing, memory_deduplication
- already-clean, added free: memory_observation, sqlite_manager

Verified: `mypy <29 files> --ignore-missing-imports --no-strict-optional
--follow-imports=silent --explicit-package-bases` -> "Success: no issues
found in 29 source files". All 17 edited modules py_compile clean.

Remaining full-repo mypy debt drops 898/90 -> 871/73 files; the full-tree
mypy stays informational until that backlog is annotated.

Authored By: Vincent S. Pereira <vincentspereira@outlook.com>

+++ src/memory_importance_scorer.py
@@ -80,7 +80,7 @@ def _component_recency(last_accessed_at: Optional[Any]) -> float:
         if last_accessed_at.tzinfo is None:
             last_accessed_at = last_accessed_at.replace(tzinfo=UTC)
         delta_days = (now - last_accessed_at).total_seconds() / 86400.0
-        return max(0.0, 1.0 - delta_days / _MAX_AGE_DAYS)
+        return float(max(0.0, 1.0 - delta_days / _MAX_AGE_DAYS))
     except Exception:
         return 0.0
